Refactor de remoção de item de carrinho

+++ src/main/java/br/com/fatec/chopperhousegames/core/domain/service/CarrinhoService.java
@@ -3,6 +3,6 @@ package br.com.fatec.chopperhousegames.core.domain.service;
 import br.com.fatec.chopperhousegames.core.domain.entity.Cliente;
 
 public interface CarrinhoService {
-     void adicionarItemCarrinho(Cliente cliente, Integer idJogo, Integer quantidade);
-    void removerItemCarrinho(Cliente cliente, Integer itemId);
+     void adicionarItemCarrinho(Cliente cliente, Long idJogo, Integer quantidade);
+    void removerItemCarrinho(Cliente cliente, Long itemId);
 }
